+++ Packages/Syntax/Sources/Syntax/TreeSitter/LanguageRegistry.swift
@@ -31,6 +31,7 @@ import SwiftTreeSitter
 import TreeSitterCSS
 import TreeSitterHTML
 import TreeSitterJavaScript
+import TreeSitterRuby
 import TreeSitterSwift
 
 public final class LanguageRegistry: Sendable {
@@ -41,6 +42,7 @@ public final class LanguageRegistry: Sendable {
         case css = "CSS"
         case html = "HTML"
         case javaScript = "JavaScript"
+        case ruby = "Ruby"
         case swift = "Swift"
         
         
@@ -134,6 +136,7 @@ private extension LanguageRegistry.Language {
             case "css": self = .css
             case "html": self = .html
             case "javascript": self = .javaScript
+            case "ruby": self = .ruby
             case "swift": self = .swift
             default: return nil
         }
@@ -147,6 +150,7 @@ private extension LanguageRegistry.Language {
             case .css: "css"
             case .html: "html"
             case .javaScript: "javascript"
+            case .ruby: "ruby"
             case .swift: "swift"
         }
     }
@@ -158,6 +162,7 @@ private extension LanguageRegistry.Language {
             case .css: unsafe tree_sitter_css()
             case .html: unsafe tree_sitter_html()
             case .javaScript: unsafe tree_sitter_javascript()
+            case .ruby: unsafe tree_sitter_ruby()
             case .swift: unsafe tree_sitter_swift()
         }
     }